Improve comments
